+++ src/server/trpc/router/organization-sign-up-form.ts
@@ -1,5 +1,5 @@
 import { enforceOrganizationAdmin } from "@/server/common/organization-perms";
-import { alphanumericNanoid } from "@/utils/alphanumericNanoid";
+import { alphanumericNanoid } from "@/utils/alphanumeric-nanoid";
 import { SignUpFieldType } from "@prisma/client";
 import { TRPCError } from "@trpc/server";
 import { nanoid } from "nanoid";
